# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Interview API
  version: "v1"
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
paths:
  /api/v1/auth/register:
    post:
      summary: 注册账号
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
  /api/v1/auth/login:
    post:
      summary: 登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /api/v1/auth/refresh:
    post:
      summary: 刷新令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        "200":
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /api/v1/auth/logout:
    post:
      summary: 登出
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        "204":
          description: 登出成功
  /api/v1/auth/forgot-password:
    post:
      summary: 忘记密码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        "202":
          description: 已受理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
  /api/v1/auth/reset-password:
    post:
      summary: 重置密码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        "204":
          description: 重置成功
  /api/v1/auth/me:
    get:
      summary: 获取当前用户
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
  /api/v1/resumes:
    get:
      summary: 查询简历列表
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeListResponse'
    post:
      summary: 上传简历
      parameters:
        - in: header
          name: X-Idempotency-Key
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: 成功
  /api/v1/resumes/{resume_id}:
    delete:
      summary: 删除简历
      parameters:
        - in: path
          name: resume_id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: 删除成功
  /api/v1/interviews:
    post:
      summary: 创建面试会话
      parameters:
        - in: header
          name: X-Idempotency-Key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterviewCreateRequest'
      responses:
        "200":
          description: 成功
  /api/v1/interviews/{interview_id}/status:
    get:
      summary: 查询面试会话状态
      parameters:
        - in: path
          name: interview_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 成功
  /api/v1/interviews/{interview_id}/playback:
    get:
      summary: 查询面试回放详情
      parameters:
        - in: path
          name: interview_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewPlaybackResponse'
  /api/v1/interviews/{interview_id}/turns:
    get:
      summary: 查询面试轮次列表
      parameters:
        - in: path
          name: interview_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  interview_id: { type: string }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        turn_id: { type: string }
                        interview_id: { type: string }
                        stage: { type: string }
                        answer_text: { type: string }
                        next_question: { type: string }
                        live_score: { type: integer }
                        generation_mode: { type: string }
                        input_source: { type: string, nullable: true }
                        asr_provider: { type: string, nullable: true }
                        llm_provider: { type: string, nullable: true }
                        tts_provider: { type: string, nullable: true }
                        degrade_flags:
                          type: array
                          items: { type: string }
                        trace_id: { type: string, nullable: true }
                        latency_ms: { type: integer }
                        created_at: { type: string }
        "404":
          description: 会话不存在
    post:
      summary: 提交面试轮次
      parameters:
        - in: path
          name: interview_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stage]
              properties:
                stage: { type: string }
                answer_text: { type: string }
                asr_text: { type: string }
                answer_audio_url: { type: string }
                answer_audio_format: { type: string, default: mp3 }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  interview_id: { type: string }
                  stage: { type: string }
                  next_question: { type: string }
                  follow_up_count: { type: integer }
                  live_score: { type: integer }
                  output_mode: { type: string }
                  tts_audio_url: { type: string, nullable: true }
                  pipeline_meta:
                    type: object
                    properties:
                      input_source: { type: string }
                      providers:
                        type: object
                        properties:
                          asr: { type: string, nullable: true }
                          llm: { type: string, nullable: true }
                          tts: { type: string, nullable: true }
                      provider_status:
                        type: object
                        properties:
                          asr: { type: string }
                          llm: { type: string }
                          tts: { type: string }
                      degrade_flags:
                        type: array
                        items: { type: string }
                      trace_id: { type: string }
                      latency_ms: { type: integer }
                      generation_mode: { type: string, enum: [local_ai, fallback_template, mock] }
        "409":
          description: 状态冲突
        "502":
          description: 上游服务失败
        "504":
          description: 上游超时
  /api/v1/interviews/{interview_id}/turns/audio:
    post:
      summary: 上传音频并提交面试轮次
      parameters:
        - in: path
          name: interview_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [stage, file]
              properties:
                stage: { type: string }
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 成功
  /api/v1/interviews/{interview_id}/finish:
    post:
      summary: 结束面试并触发报告
      parameters:
        - in: path
          name: interview_id
          required: true
          schema: { type: string }
      responses:
        "202":
          description: 已接收
  /api/v1/report/{interview_id}:
    get:
      summary: 查询面试报告
      parameters:
        - in: path
          name: interview_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 成功
  /api/v1/report/{interview_id}/retry:
    post:
      summary: 重试报告生成
      parameters:
        - in: path
          name: interview_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 已触发重试
  /api/v1/interviews/history:
    get:
      summary: 查询历史记录
      parameters:
        - in: query
          name: job_role
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: 成功
  /api/v1/admin/imports/materials:
    post:
      summary: 触发材料导入异步任务
      parameters:
        - in: header
          name: X-Idempotency-Key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialImportRequest'
      responses:
        "202":
          description: 已受理
  /api/v1/admin/imports/materials/{task_id}:
    get:
      summary: 查询材料导入任务状态
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 成功
        "403":
          description: 权限不足
  /api/v1/admin/providers/health:
    get:
      summary: 查询 provider 健康状态
      description: 健康结果基于 SDK 初始化与最小调用，不代表外部 HTTP 服务可达性。
      security: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  overall: { type: string, enum: [UP, DOWN, DEGRADED] }
                  providers:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        status: { type: string, enum: [UP, DOWN, DEGRADED] }
                        provider: { type: string }
                        model: { type: string }
                        latency_ms: { type: integer }
                        error_message: { type: string }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        display_name: { type: string }
    RegisterResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/AuthUserProfile'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TokenRefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ForgotPasswordResponse:
      type: object
      properties:
        accepted: { type: boolean }
        message: { type: string }
    ResetPasswordRequest:
      type: object
      required: [reset_token, new_password]
      properties:
        reset_token: { type: string }
        new_password: { type: string, minLength: 8 }
    AuthUserProfile:
      type: object
      properties:
        user_id: { type: string }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [user, admin] }
        status: { type: string, enum: [active, disabled] }
    AuthTokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [bearer] }
        expires_in: { type: integer }
        refresh_token: { type: string }
        user:
          $ref: '#/components/schemas/AuthUserProfile'
    AuthMeResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/AuthUserProfile'
    ResumeListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              resume_id: { type: string }
              file_name: { type: string }
              parse_status: { type: string }
              created_at: { type: string }
              last_used_at: { type: string, nullable: true }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    InterviewCreateRequest:
      type: object
      required: [resume_id, job_role]
      properties:
        resume_id: { type: string }
        job_role: { type: string, enum: [java, web] }
        difficulty: { type: string, enum: [easy, medium, hard] }
        input_mode: { type: string, enum: [text, voice] }
        output_mode: { type: string, enum: [text, voice] }
    InterviewTurnRequest:
      type: object
      required: [stage]
      properties:
        stage: { type: string }
        answer_text: { type: string }
        asr_text: { type: string }
    InterviewPlaybackResponse:
      type: object
      properties:
        interview_id: { type: string }
        resume:
          type: object
          properties:
            resume_id: { type: string }
            file_name: { type: string }
        meta:
          type: object
          properties:
            job_role: { type: string }
            difficulty: { type: string }
            status: { type: string }
            started_at: { type: string }
            finished_at: { type: string, nullable: true }
        turns:
          type: array
          items:
            type: object
            properties:
              turn_id: { type: string }
              sequence: { type: integer }
              question: { type: string }
              answer: { type: string }
              question_ts: { type: string }
              answer_ts: { type: string, nullable: true }
    MaterialImportRequest:
      type: object
      properties:
        rebuild_mode: { type: string, enum: [full, incremental], default: full }
        roles:
          type: array
          items: { type: string, enum: [java, web] }
        dry_run: { type: boolean, default: false }
        chunk_model: { type: string, default: qwen2.5:7b }
        embedding_model: { type: string, default: nomic-embed-text }